requirements.txt
pandas>=2.1.0
numpy>=1.25.2
PyYAML>=6.0.1
